demo-deepagents/pyproject.toml
[project]
name = "demo-deepagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents",
    "langchain-openrouter",
    "prompt-toolkit>=3.0.52",
    "python-dotenv",
    "rich>=14.3.3",
]
